+++ packages/terra-modal/tests/nightwatch/components/ModalCustomProps.jsx
@@ -1,0 +1,49 @@
+import React from 'react';
+import Modal from '../../../lib/Modal';
+
+class ModalCustomProps extends React.Component {
+  constructor() {
+    super();
+
+    this.state = {
+      isOpen: true,
+    };
+
+    this.handleOpenModal = this.handleOpenModal.bind(this);
+    this.handleCloseModal = this.handleCloseModal.bind(this);
+  }
+
+
+  handleOpenModal() {
+    this.setState({ isOpen: true });
+  }
+
+  handleCloseModal() {
+    this.setState({ isOpen: false });
+  }
+
+  render() {
+    return (
+      <div>
+        <Modal
+          id="custom-props"
+          ariaLabel="Terra Modal"
+          isOpen={this.state.isOpen}
+          onRequestClose={this.handleCloseModal}
+        >
+          <div>
+            <h1>Terra Modal</h1>
+            <h2>Subtitle</h2>
+            <hr />
+            <p>The Terra Modal is appended to the document body.</p>
+            <p>{'Modal is assigned a role of \'document\' for accessibility.'}</p>
+            <button onClick={this.handleCloseModal}>Close Modal</button>
+          </div>
+        </Modal>
+        <button className="button-open-modal" onClick={this.handleOpenModal}>Open Modal</button>
+      </div>
+    );
+  }
+}
+
+export default ModalCustomProps;